+++ tests/test_wellknowntypes.py
@@ -84,7 +84,6 @@ class TestWellKnownTypes(pylint.testutils.CheckerTestCase):
             with self.assertNoMessages():
                 self.walk(node.root())
 
-    @pytest.mark.xfail(reason='not actually checking fields on well-known types')
     @pytest.mark.parametrize("module,wkt,_", SAMPLE_WKTS)
     def test_wkt_should_still_warn(self, module, wkt, _, error_on_missing_modules):
         node = astroid.extract_node("""
@@ -93,6 +92,16 @@ class TestWellKnownTypes(pylint.testutils.CheckerTestCase):
             t.should_warn = 123
         """.format(module=module, wkt=wkt))
         msg = make_message(node.targets[0], wkt, 'should_warn')
+        with self.assertAddsMessages(msg):
+            self.walk(node.root())
+
+    @pytest.mark.parametrize("module,wkt,_", SAMPLE_WKTS)
+    def test_wkt_kwargs_should_still_warn(self, module, wkt, _, error_on_missing_modules):
+        node = astroid.extract_node("""
+            from google.protobuf.{module} import {wkt}
+            {wkt}(should_warn=123)
+        """.format(module=module, wkt=wkt))
+        msg = pylint.testutils.Message('unexpected-keyword-arg', node=node, args=('should_warn', 'constructor'))
         with self.assertAddsMessages(msg):
             self.walk(node.root())