+++ CHANGELOG.md
@@ -17,6 +17,8 @@
   Thanks @zoedberg for the report
 - Update handling of messages with full names matching well-known types to
   match protobuf behaviour (#45). Thanks @zwn for the report
+- Remove special-case handling of well-known types and update to match
+  existing behaviour of other protobuf messages
 
 ## [0.19.0] - 2021-03-24
 - Add support for imported message definitions via 'import "foo.proto";'